Split the lifecycle and interaction blocks into seven extension files
The next contiguous stretch moves verbatim: RuntimeAndLifecycle (bootstrap,
runtime and proxy recovery, shutdown, notifications, metadata refresh),
SessionLifecycle (clear/new/resume/worker switching), ChatOrganization
(folders, pin, order, archive, delete with undo), SplitChatPanes (with the
chat-export entry points that share the pane machinery), Workspaces,
ComposerAttachments (context pack, attachments, live app and simulator
capture), and PlanAndCheckpoints (checkpoints, rewind, plan approval,
question resolution). Shared private members widen with the breadcrumb.

Behavior-identical file split; the full suite passes unchanged.
